+++ hyraigne/src/sites/mangadex/page.rs
@@ -1,6 +1,9 @@
 use super::{
     chapter,
-    selectors::PAGE_URL_SELECTOR,
+    models::{
+        ChapterDetail,
+        Response,
+    },
 };
 use crate::{
     Chapter,
@@ -8,51 +11,54 @@ use crate::{
     Page,
     Result,
 };
-use kuchiki::traits::*;
 use std::path::{
     Path,
     PathBuf,
 };
 use url::Url;
 
-/// Scrape page links from the chapter's page HTML.
-#[allow(clippy::filter_map)]
-pub(super) fn scrape_from_html<'a>(
-    html: &kuchiki::NodeRef,
+/// Extract page links from the API response.
+pub(super) fn extract_from_response<'a>(
+    response: &Response<ChapterDetail>,
     chapter: &'a Chapter<'_>,
 ) -> Result<Vec<Page<'a>>> {
-    PAGE_URL_SELECTOR
-        .filter(html.descendants().elements())
-        .map(|node| {
-            let url = node
-                .attributes
-                .borrow()
-                .get("data-url")
-                .ok_or_else(|| {
-                    Error::Scraping("page URL not found".to_owned())
-                })?
-                .to_owned();
-
-            let url = Url::parse(&url).map_err(|err| {
-                Error::Scraping(format!("invalid page URL `{}`: {}", url, err))
-            })?;
+    response
+        .data
+        .pages
+        .iter()
+        .map(|page| {
+            let server_url = &response.data.server;
+            let fallback_url = &response.data.server_fallback;
+            let path = format!("{}/{}", response.data.hash, page);
 
             Ok(Page {
                 chapter,
-                main: url,
-                fallback: None,
+                main: urljoin(server_url, &path)?,
+                fallback: Some(urljoin(fallback_url, &path)?),
             })
         })
-        .collect::<Result<Vec<_>>>()
+        .collect()
 }
 
 /// Get the file path of the page on disk.
 pub(super) fn get_path(path: &Path, page: &Page<'_>, index: usize) -> PathBuf {
     let dirpath = chapter::get_path(path, page.chapter);
     let extension = crate::fs::extname_from_url(&page.main);
-    let filename = format!("{:03}.{}", index, extension);
+    let filename = format!("{:03}-{:03}.{}", page.chapter.id, index, extension);
 
     [&dirpath, Path::new(&filename)].iter().collect()
+}
+
+/// Append `suffix` to `base` and parse the result as an URL.
+fn urljoin(base: &Url, suffix: &str) -> Result<Url> {
+    base.join(&suffix).map_err(|err| {
+        Error::Scraping(format!(
+            "failed to join {} with {}: {}",
+            base.as_str(),
+            suffix,
+            err
+        ))
+    })
 }
 
 // Tests {{{
@@ -70,20 +76,20 @@ mod tests {
         let series = Series {
             title: "Example".to_owned(),
             url: Url::parse("http://example.com/").unwrap(),
-            pagination: Pagination::new(78, 10),
+            pagination: Pagination::new(0, 0),
         };
         let chapter = Chapter {
-            id: 10.0,
+            id: 30.0,
             series: &series,
-            volume: None,
-            url: Url::parse("http://example.com/10/").unwrap(),
+            volume: Some("10".to_owned()),
+            url: Url::parse("http://example.com/30/").unwrap(),
         };
         let page = Page {
             chapter: &chapter,
             main: Url::parse("http://example.com/10/uWu.jpg").unwrap(),
             fallback: None,
         };
-        let expected = "Downloads/Example/Example 010/042.jpg";
+        let expected = "Downloads/Example/Example 10/030-042.jpg";
 
         let path = get_path(Path::new("Downloads"), &page, 42);
 
@@ -95,7 +101,7 @@ mod tests {
         let series = Series {
             title: "Example".to_owned(),
             url: Url::parse("http://example.com/").unwrap(),
-            pagination: Pagination::new(1, 0),
+            pagination: Pagination::new(0, 0),
         };
         let chapter = Chapter {
             id: 10.0,
@@ -104,13 +110,13 @@ mod tests {
             url: Url::parse("http://example.com/10/").unwrap(),
         };
         let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-        path.push("testdata/webtoons.com/chapter.html");
-        let html = std::fs::read_to_string(&path).expect("test data");
-        let document = kuchiki::parse_html().one(html);
+        path.push("testdata/mangadex.org/chapter.json");
+        let json = std::fs::read_to_string(&path).expect("test data");
+        let response = serde_json::from_str(&json).expect("invalid JSON");
 
-        let pages = scrape_from_html(&document, &chapter).unwrap();
+        let pages = extract_from_response(&response, &chapter).unwrap();
 
-        assert_eq!(pages.len(), 32);
+        assert_eq!(pages.len(), 62);
     }
 }